requirements.txt
pymupdf==1.26.3
fastapi==0.116.1
genanki==0.13.1
openai==1.96.1
Pillow==11.3.0
pytesseract==0.3.13
python-decouple==3.8
python_docx==1.1.2
tiktoken==0.7.0
Django==5.0.4                 # web framework
djangorestframework==3.16.0   # JSON API layer (DRF)